pyproject.toml
[tool.poetry]
name = "habits-drf"
version = "0.1.0"
description = ""
authors = ["DimaSimakov <dimarodeo@gmail.com>"]
readme = "README.md"
packages = [{include = "habits_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
